+++ mantisshrimp/utils/__init__.py
@@ -2,5 +2,5 @@ from mantisshrimp.utils.utils import *
 from mantisshrimp.utils.torch_utils import *
 from mantisshrimp.utils.imageio import *
 from mantisshrimp.utils.get_files import *
-from mantisshrimp.utils.download_url import *
+from mantisshrimp.utils.download_utils import *
 from mantisshrimp.utils.data_dir import *
